Conflicts:
	example/advanced/cloud_trace/cloud_data.cpp
	example/advanced/spectra/shared_objects.cpp
	example/oglplus/021_overdraw.cpp
	example/oglplus/022_vase.cpp
	example/oglplus/027_brain_scan_vis.cpp
	example/oglplus/030_pin_display.cpp
	example/standalone/001_gl_limits.cpp
	implement/oglplus/enums/ext/nv_path_command_names.ipp
	implement/oglplus/images/cloud.ipp
- Also updated the eagine submodule.

+++ example/advanced/spectra/calculator_gpu.cpp
@@ -4,7 +4,7 @@
  *
  *  @author Matus Chochlik
  *
- *  Copyright 2012-2014 Matus Chochlik. Distributed under the Boost
+ *  Copyright 2012-2015 Matus Chochlik. Distributed under the Boost
  *  Software License, Version 1.0. (See accompanying file
  *  LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
  */
@@ -219,7 +219,7 @@ void SpectraDefaultGPUMatrixTransf::BeginBatch(void)
 	input_tex.Bind(Texture::Target::Buffer);
 	prog_input_data.Set(1);
 
-	prog_input_size.Set(in_size);
+	prog_input_size.Set(int(in_size));
 
 	vao.Bind();
 
@@ -257,7 +257,7 @@ unsigned SpectraDefaultGPUMatrixTransf::BeginTransform(
 		input
 	);
 
-	prog_input_offs.Set(current_transform*in_size);
+	prog_input_offs.Set(int(current_transform*in_size));
 
 	xfbs[current_transform].Bind();
