+++ example/oglplus/031_sketch.cpp
@@ -598,10 +598,10 @@ public:
 		);
 
 		// setup the camera
-		Vec3f camera_target(0.0, 0.8, 0.0);
+		Vec3f camera_target(0.0f, 0.8f, 0.0f);
 		auto camera = CamMatrixf::Orbiting(
 			camera_target,
-			7.0 - SineWave(time / 14.0)*3.0,
+			GLfloat(7.0 - SineWave(time / 14.0)*3.0),
 			FullCircles(time / 26.0),
 			Degrees(45 + SineWave(time / 17.0) * 40)
 		);
