+++ example/standalone/004_any_text_rendering.cpp
@@ -132,7 +132,7 @@ public:
 		rndr.SetCamera(
 			CamMatrixf::Orbiting(
 				Vec3f(),
-				9.0 + SineWave(t / 7.0)*3.0,
+				GLfloat(9.0 + SineWave(t / 7.0)*3.0),
 				-FullCircles(t / 17.0),
 				Degrees(SineWave(t / 21.0) * 35)
 			)
